Export: split AssetWriter into video audio and lifecycle units
